+++ src/python/pants/core/util_rules/partitions.py
